+++ packages/ooxml-common/src/lib.rs
@@ -7,3 +7,4 @@
 
 pub mod chart;
 pub mod color;
+pub mod zip;